requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.24.1 